+++ packages/examples/core/di/ts/contentChild/module.ts
@@ -6,15 +6,6 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {NgModule} from '@angular/core';
-import {BrowserModule} from '@angular/platform-browser';
-import {ContentChildComp, Pane, Tab} from './content_child_example';
-
-@NgModule({
-  imports: [BrowserModule],
-  declarations: [ContentChildComp, Pane, Tab],
-  bootstrap: [ContentChildComp],
-})
-export class AppModule {}
+import {ContentChildComp} from './content_child_example';
 
 export {ContentChildComp as AppComponent};
