+++ packages/examples/core/di/ts/contentChildren/content_children_howto.ts
@@ -11,13 +11,11 @@ import {AfterContentInit, ContentChildren, Directive, QueryList} from '@angular/
 
 @Directive({
   selector: 'child-directive',
-  standalone: false,
 })
 class ChildDirective {}
 
 @Directive({
   selector: 'someDir',
-  standalone: false,
 })
 class SomeDir implements AfterContentInit {
   @ContentChildren(ChildDirective) contentChildren!: QueryList<ChildDirective>;
