+++ packages/examples/core/di/ts/contentChildren/module.ts
@@ -6,15 +6,6 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {NgModule} from '@angular/core';
-import {BrowserModule} from '@angular/platform-browser';
-import {ContentChildrenComp, Pane, Tab} from './content_children_example';
-
-@NgModule({
-  imports: [BrowserModule],
-  declarations: [ContentChildrenComp, Pane, Tab],
-  bootstrap: [ContentChildrenComp],
-})
-export class AppModule {}
+import {ContentChildrenComp} from './content_children_example';
 
 export {ContentChildrenComp as AppComponent};
