+++ packages/examples/core/di/ts/viewChild/module.ts
@@ -6,16 +6,6 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {NgModule} from '@angular/core';
-import {BrowserModule} from '@angular/platform-browser';
-
-import {Pane, ViewChildComp} from './view_child_example';
-
-@NgModule({
-  imports: [BrowserModule],
-  declarations: [ViewChildComp, Pane],
-  bootstrap: [ViewChildComp],
-})
-export class AppModule {}
+import {ViewChildComp} from './view_child_example';
 
 export {ViewChildComp as AppComponent};
