+++ packages/examples/core/di/ts/viewChildren/module.ts
@@ -6,16 +6,6 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {NgModule} from '@angular/core';
-import {BrowserModule} from '@angular/platform-browser';
-
-import {Pane, ViewChildrenComp} from './view_children_example';
-
-@NgModule({
-  imports: [BrowserModule],
-  declarations: [ViewChildrenComp, Pane],
-  bootstrap: [ViewChildrenComp],
-})
-export class AppModule {}
+import {ViewChildrenComp} from './view_children_example';
 
 export {ViewChildrenComp as AppComponent};
