+++ packages/examples/core/testability/ts/whenStable/module.ts
@@ -5,4 +5,4 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.dev/license
  */
-export {AppModule, StableTestCmp as AppComponent} from './testability_example';
+export {StableTestCmp as AppComponent} from './testability_example';
