+++ packages/examples/core/testability/ts/whenStable/testability_example.ts
@@ -6,8 +6,7 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {Component, NgModule} from '@angular/core';
-import {BrowserModule} from '@angular/platform-browser';
+import {Component} from '@angular/core';
 
 @Component({
   selector: 'example-app',
@@ -15,7 +14,6 @@ import {BrowserModule} from '@angular/platform-browser';
     <button class="start-button" (click)="start()">Start long-running task</button>
     <div class="status">Status: {{ status }}</div>
   `,
-  standalone: false,
 })
 export class StableTestCmp {
   status = 'none';
@@ -26,6 +24,3 @@ export class StableTestCmp {
     }, 5000);
   }
 }
-
-@NgModule({imports: [BrowserModule], declarations: [StableTestCmp], bootstrap: [StableTestCmp]})
-export class AppModule {}
