+++ packages/examples/core/ts/metadata/encapsulation.ts
@@ -30,7 +30,6 @@ import {Component, ViewEncapsulation} from '@angular/core';
     `,
   ],
   encapsulation: ViewEncapsulation.ShadowDom,
-  standalone: false,
 })
-class MyApp {}
+export class MyApp {}
 // #enddocregion
