+++ packages/examples/core/ts/pipes/truncate.ts
@@ -10,7 +10,6 @@ import {Pipe, PipeTransform} from '@angular/core';
 
 @Pipe({
   name: 'truncate',
-  standalone: false,
 })
 export class TruncatePipe implements PipeTransform {
   transform(value: string, length: number, symbol: string) {
