+++ packages/examples/core/ts/prod_mode/my_component.ts
@@ -11,6 +11,5 @@ import {Component} from '@angular/core';
 @Component({
   selector: 'my-component',
   template: '<h1>My Component</h1>',
-  standalone: false,
 })
 export class MyComponent {}
