+++ packages/examples/core/ts/prod_mode/prod_mode_example.ts
@@ -6,14 +6,17 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-import {enableProdMode, NgModule} from '@angular/core';
-import {BrowserModule, platformBrowser} from '@angular/platform-browser';
-
+import {enableProdMode} from '@angular/core';
+import {bootstrapApplication} from '@angular/platform-browser';
 import {MyComponent} from './my_component';
 
 enableProdMode();
 
-@NgModule({imports: [BrowserModule], declarations: [MyComponent], bootstrap: [MyComponent]})
-export class AppModule {}
-
-platformBrowser().bootstrapModule(AppModule);
+bootstrapApplication(MyComponent, {
+  providers: [
+    // Add your global providers here
+    // provideRouter(routes),
+    // provideHttpClient(),
+    // etc.
+  ],
+});
